requirements.txt
appdirs==1.4.4
asgiref==3.7.2
asttokens==2.2.1
backcall==0.2.0
black==23.3.0
click==8.1.3
decorator==5.1.1
Django==4.2.2
djangorestframework==3.14.0
executing==1.2.0
ipdb==0.13.13
ipython==8.14.0
jedi==0.18.2
matplotlib-inline==0.1.6
mypy-extensions==1.0.0
packaging==23.1
parso==0.8.3
pathspec==0.11.1
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.6.0
prompt-toolkit==3.0.38
ptpython==3.0.23
ptyprocess==0.7.0
pure-eval==0.2.2
pycryptodome==3.18.0
Pygments==2.15.1
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
traitlets==5.9.0
wcwidth==0.2.6
